Refactor: move buildApplicationEndpoint from container to dao
